Move view components into own folder
