+++ examples/recipes/src/request_response/RequestAuditLog.tsx
@@ -1,5 +1,5 @@
 import { useState } from 'react'
-import { RequestAuditLogView } from './RequestAuditLogView'
+import { RequestAuditLogView } from './components/RequestAuditLogView'
 import { useElectricQueryLog, PaginationState } from './use_electric_query_log'
 
 export const RequestAuditLog = () => {
@@ -11,8 +11,7 @@ export const RequestAuditLog = () => {
 
   const { requestLogs, totalNumberOfRequests } = useElectricQueryLog({
     pagination,
-    startDate: new Date('2024-02-20 11:00:00.000Z'),
-    endDate: new Date('2024-02-21 04:05:00.000Z'),
+    // can also specify datetime range
   })
 
   return (
