+++ examples/recipes/src/request_response/components/RequestResultView.tsx
